requirements.txt
anthropic>=0.40.0
apify-client>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
requests>=2.32.0
pytest>=8.0.0
pytest-cov>=4.1.0
